enh: Crop/pad using only mask if set, 'Dirichlet boundary condition' [Registration]
By using only the domain mask to set the CP status, the FFD cropping/padding is independent of the actual image content and thus will be identical for all pairwise registrations performed with the same domain mask.

When the images are cropped, but not the FFD lattice, extend the registration domain by two times the FFD kernel radius for velocity based FFDs. When the FFD lattice is also cropped, only extend domain by one times the kernel radius.

The ‘Dirichlet boundary condition’ forces the parameters of CPs at the boundary of the finite domain to be zero. This ensures that the velocity of a SV FFD is zero at the boundary.